dashboard_api/requirements.txt
alembic>=0.7
attrs>=21.4.0
click>=8.0
flasgger>=0.9.5
Flask==2.1.0
flask-cors==3.0.10
Flask-Migrate==2.5.2
flask-restplus>=0.13.0
Flask-Script==2.0.6
flask-sqlalchemy==2.5.1
gunicorn >= 20.1.0
importlib-metadata>=3.6.0
itsdangerous>=2.0
Jinja2>=3.0
jsonschema>=4.6.2
Mako
MarkupSafe>=2.0
mistune>=2.0.3
numpy==1.21.5
pandas==1.3.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyrsistent>=0.18.1
python-dateutil==2.8.2
python-dotenv>=0.15.0
pytz
PyYAML>=6.0
SQLAlchemy==1.4.0
Werkzeug==2.0
wheel >= 0.37.1
zipp>=0.5
